+++ app/api/auth/index.js
@@ -3,4 +3,3 @@ import captchaAuthorization from './captchaMiddleware.js';
 import { CaptchaModel } from './CaptchaModel.js';
 
 export { needsAuthorization, captchaAuthorization, CaptchaModel };
-export { validatePasswordMiddleWare } from './validatePasswordMiddleWare.js';
