+++ app/api/core/infrastructure/express/users/specs/DeleteUserController.spec.ts
@@ -8,7 +8,7 @@ import { userRoutes } from '../routes.js';
 import { fixtures, f } from './fixtures.js';
 import { UserGroupDBO } from '#api/core/infrastructure/mongodb/user/UserGroupDBO.js';
 
-jest.mock('../../../../../auth/validatePasswordMiddleWare.ts', () => ({
+jest.mock('../ValidatePasswordMiddleWare.ts', () => ({
   validatePasswordMiddleWare: (_req: Request, _res: Response, next: NextFunction) => next(),
 }));
