+++ app/api/core/infrastructure/express/users/specs/GetUsersController.spec.ts
@@ -6,7 +6,7 @@ import { setUpApp } from '#api/utils/testingRoutes.js';
 import { userRoutes } from '../routes.js';
 import { fixtures, f } from './fixtures.js';
 
-jest.mock('../../../../../auth/validatePasswordMiddleWare.ts', () => ({
+jest.mock('../ValidatePasswordMiddleWare.ts', () => ({
   validatePasswordMiddleWare: (_req: Request, _res: Response, next: NextFunction) => next(),
 }));
