+++ app/api/core/infrastructure/express/users/specs/ResetPasswordController.spec.ts
@@ -7,7 +7,7 @@ import { userRoutes } from '../routes.js';
 import { fixtures, f } from './fixtures.js';
 import { UserRole } from '#shared/types/userSchema.js';
 
-jest.mock('../../../../../auth/validatePasswordMiddleWare.ts', () => ({
+jest.mock('../ValidatePasswordMiddleWare.ts', () => ({
   validatePasswordMiddleWare: (_req: Request, _res: Response, next: NextFunction) => next(),
 }));
